gitlab-ci: Run jobs on the gcc:9 image.
Use a versioned gcc container image for more consistent
test results. This is the same version we're using for
other projects, but of course it will need to be bumped
periodically.

The current gcc release is 10.2. The oldest supported
release is 8.4, so this is in the middle of the support
window.

Signed-off-by: Marvin Scholz <epirat07@gmail.com>